+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2021-10-05
 - Made the decision to write the front-end with MobX + React.
-- Started working on a front-end refactor for MobX + React under `src/fex` 
+- Started working on a front-end refactor for MobX + React under `src/fex`
+    - Implemented connection status and active user count using purely MobX + React.
 # 2021-10-04
 - Created class for a Challenge and a User in the back-end.
 - Created abstract class for game states (`game_state`)